feat(connect): split `composeTransaction` and `sendTransaction`
