+++ packages/connect-common/src/types/api/bitcoin/index.ts
@@ -4,11 +4,13 @@ import { Type } from '@trezor/schema-utils';
 import type { authorizeCoinjoin } from './authorizeCoinjoin';
 import type { cancelCoinjoinAuthorization } from './cancelCoinjoinAuthorization';
 import type { composeTransaction } from './composeTransaction';
+import type { sendTransaction } from './sendTransaction';
 import type { signTransaction } from './signTransaction';
 
 // Bitcoin-specific operations
 export const TrezorConnectBitcoin = Type.Object({
     signTransaction: Type.Unsafe<typeof signTransaction>(),
+    sendTransaction: Type.Unsafe<typeof sendTransaction>(),
     composeTransaction: Type.Unsafe<typeof composeTransaction>(),
     authorizeCoinjoin: Type.Unsafe<typeof authorizeCoinjoin>(),
     cancelCoinjoinAuthorization: Type.Unsafe<typeof cancelCoinjoinAuthorization>(),
