+++ packages/connect-common/src/types/api/bitcoin/sendTransaction.ts
@@ -1,0 +1,6 @@
+import type { ComposeParams, SignedTransaction } from './common';
+import type { Params, Response } from '../../params';
+
+type SendParams = ComposeParams & { push?: boolean; identity?: string };
+
+export declare function sendTransaction(params: Params<SendParams>): Response<SignedTransaction>;
