+++ suite-common/wallet-core/src/send/sendFormBitcoinThunks.ts
@@ -118,7 +118,6 @@ export const composeBitcoinTransactionFeeLevelsThunk = createThunk<
             : account.addresses.change;
 
         const params: Parameters<typeof TrezorConnect.composeTransaction>[0] = {
-            // needs to be present in order to correct resolve of @trezor/connect params overload
             account: {
                 path: account.path,
                 addresses: {
@@ -192,7 +191,6 @@ export const composeBitcoinTransactionFeeLevelsThunk = createThunk<
                 customLevels.length > 0
                     ? await TrezorConnect.composeTransaction({
                           ...params,
-                          account: params.account, // needs to be present in order to correct resolve type of @trezor/connect params overload
                           feeLevels: customLevels,
                       })
                     : ({ success: false } as const);
